requirements.txt
fastapi
uvicorn
psycopg[binary]
python-dotenv
email-validator